[WIP] Write docker engine and finish CLI
This diff contains two major components. The first is completing the
`DockerEngine` so that it runs actual commands. To do so, we introduced
both the `exec` and `reporter` packages, which are responsible for
executing commands and logging results. We dependency inject these when
we create an engine to make unit testing easier. Actual unit testing is
left for a different diff.

Additionally, this diff finishes the CLI for VMs. All cli commands
redirect to the proper virtualhost/engine method. We just need to write
the `vagrant` engine.