Add integration testing
Add an integration test script which simulates the e2e usage of a
container tmplinux. We can run this script both locally and on
travis-ci, because travis-ci supports docker. We cannot write automated
integration tests for vm, because travis-ci doesn't support any vm
platforms.

Additionally, remove unnecessary unit test files. We do not include unit
tests for packages in which all of the tests would be trivial.