Add VagrantEngine for vm
Add the Vagrant Engine to drive vms. We can now create virtual linux
environments.

One choice I made here was to use the directory to control the Vagrant
environment, instead of the global id. I'm not sure if thats a good
decision - I can revisit it later if I want.